重构知识库功能，为rwkv支持做准备
