+++ views/index.html
@@ -106,7 +106,7 @@
 				<v-app-bar-nav-icon @click.stop="drawer = !drawer"></v-app-bar-nav-icon>
 				<v-toolbar-title>闻达</v-toolbar-title>
 				<v-spacer></v-spacer>
-				<v-switch hide-details color="purple" inset v-model="chatPDF" label="chatPDF"></v-switch>
+				<v-switch hide-details color="purple" inset v-model="zhishiku" label="知识库"></v-switch>
 				&nbsp;&nbsp;&nbsp;&nbsp;
 				<v-hover v-slot="{ hover }">
 					<div>
